+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/handler/AnnotationHandlerMappingTest.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
